[WIP] Platform tests
