+++ rest/types.go
@@ -51,9 +51,9 @@ type Platform struct {
 	ID          string       `json:"id"`
 	Type        string       `json:"type"`
 	Name        string       `json:"name"`
-	Description string       `json:"description,omitempty"`
-	CreatedAt   time.Time    `json:"created_at,omitempty"`
-	UpdatedAt   time.Time    `json:"updated_at,omitempty"`
+	Description string       `json:"description"`
+	CreatedAt   time.Time    `json:"created_at"`
+	UpdatedAt   time.Time    `json:"updated_at"`
 	Credentials *Credentials `json:"credentials,omitempty"`
 }
